feat: rename dashboard-items and add new dishs and pictures locales properties
