+++ src/components/dashboard/dashboard-items/dishs-dashboard.component.tsx
@@ -5,41 +5,41 @@ import { TFields } from "../../form/form";
 import TableDashboard from "../table-dashboard.component";
 import { TTabProps } from "src/components/tab/tab.component";
 import {
-  createPicture,
-  deletePicture,
-  updatePicture,
-} from "src/app/actions/pictures/pictures.actions";
+  createDish,
+  deleteDish,
+  updateDish,
+} from "src/app/actions/dishs/dishs.actions";
 import { useToast } from "src/app/providers/toast.provider";
 
-type TPicture = {
+type TDish = {
   id: number;
-  pictureUrl: string;
-  frenchDescription: string;
-  englishDescription: string;
-  dishServiceId?: number | null;
+  labelFR: string;
+  labelEN: string;
+  price: number;
+  dishServiceId?: number;
   updatedAt?: Date;
   createdAt?: Date;
 };
 
 export type TTableData = TTableGeneric<
-  TPicture & {
+  TDish & {
     actions: string;
   }
 >;
 
-export type TPicturesDashboardProps = {
-  pictures: TPicture[] | undefined;
+export type TDishsDashboardProps = {
+  dishs: TDish[] | undefined;
   dishServiceId: number | undefined;
   dishServiceTitle: string | undefined;
 };
 
-const PicturesDashboard = ({
+const DishsDashboard = ({
   dishServiceTitle,
-  pictures,
+  dishs,
   dishServiceId,
   tabItems,
-}: TPicturesDashboardProps & TTabProps) => {
-  const [selectedPicture, setSelectedPicture] = useState<TPicture | undefined>(
+}: TDishsDashboardProps & TTabProps) => {
+  const [selectedDish, setSelectedDish] = useState<TDish | undefined>(
     undefined
   );
   const [modalActionForm, setModalActionForm] = useState("");
@@ -47,11 +47,11 @@ const PicturesDashboard = ({
 
   useEffect(() => {
     if (modalActionForm === "createModal") {
-      setSelectedPicture(undefined);
+      setSelectedDish(undefined);
     }
   }, [modalActionForm]);
 
-  if (!pictures) return <div>Chargement...</div>;
+  if (!dishs) return <div>Chargement...</div>;
 
   const fields: TFields[] = [
     {
@@ -59,11 +59,19 @@ const PicturesDashboard = ({
       items: [
         {
           fieldElement: "input",
-          inputType: "file",
-          placeholder: "/chemin/vers/image",
-          defaultValue: selectedPicture && selectedPicture.pictureUrl,
-          label: "Url de l'image",
-          name: "pictureUrl",
+          inputType: "text",
+          placeholder: "Rouleaux ...",
+          defaultValue: selectedDish && selectedDish.labelFR,
+          label: "Nom de l'assiette en FR",
+          name: "labelFR",
+        },
+        {
+          fieldElement: "input",
+          inputType: "text",
+          placeholder: "Rouleaux ...",
+          defaultValue: selectedDish && selectedDish.labelEN,
+          label: "Nom de l'assiette en EN",
+          name: "labelEN",
         },
       ],
     },
@@ -72,19 +80,11 @@ const PicturesDashboard = ({
       items: [
         {
           fieldElement: "input",
-          inputType: "text",
-          placeholder: "Rouleau de printemps",
-          defaultValue: selectedPicture && selectedPicture.frenchDescription,
-          label: "Description de l'image",
-          name: "frenchDescription",
-        },
-        {
-          fieldElement: "input",
-          inputType: "text",
-          placeholder: "Rouleau de printemps",
-          defaultValue: selectedPicture && selectedPicture.englishDescription,
-          label: "Description de l'image",
-          name: "englishDescription",
+          inputType: "number",
+          placeholder: "10€",
+          defaultValue: selectedDish && selectedDish.price,
+          label: "Prix de l'entrée",
+          name: "price",
         },
       ],
     },
@@ -92,7 +92,7 @@ const PicturesDashboard = ({
 
   const handleSubmitAuthor = async (formData: FormData) => {
     if (modalActionForm === "createModal" && dishServiceId) {
-      const result = await createPicture(
+      const result = await createDish(
         dishServiceId,
         formData,
         dishServiceTitle
@@ -104,9 +104,9 @@ const PicturesDashboard = ({
       }
       setModalActionForm("");
     }
-    if (modalActionForm === "editModal" && selectedPicture) {
-      const result = await updatePicture(
-        selectedPicture?.pictureUrl,
+    if (modalActionForm === "editModal" && selectedDish) {
+      const result = await updateDish(
+        selectedDish.id,
         formData,
         dishServiceTitle
       );
@@ -120,11 +120,8 @@ const PicturesDashboard = ({
   };
 
   const handleDeleteAuthor = async () => {
-    if (selectedPicture) {
-      const result = await deletePicture(
-        selectedPicture?.id,
-        selectedPicture.pictureUrl
-      );
+    if (selectedDish) {
+      const result = await deleteDish(selectedDish?.id, dishServiceTitle);
       if (result?.error) {
         showToast({ message: result?.error, type: "alert" });
       } else {
@@ -138,20 +135,20 @@ const PicturesDashboard = ({
     <TableDashboard
       tabItems={tabItems}
       title={`Les ${dishServiceTitle}`}
-      data={pictures as Array<TTableData>}
+      data={dishs as Array<TTableData>}
       fields={fields}
       columns={[
         {
-          key: "pictureUrl",
-          header: "URL de l'image",
+          key: "labelFR",
+          header: "Nom de l'assiette en FR",
         },
         {
-          key: "englishDescription",
-          header: "Description de l'image en EN",
+          key: "labelEN",
+          header: "Nom de l'assiette en EN",
         },
         {
-          key: "frenchDescription",
-          header: "Description de l'image en FR",
+          key: "price",
+          header: "Prix",
         },
         {
           key: "actions",
@@ -162,9 +159,9 @@ const PicturesDashboard = ({
       setModalActionForm={setModalActionForm}
       handleDeleteModalAction={handleDeleteAuthor}
       handleCreatEditModalFormAction={handleSubmitAuthor}
-      selectedItem={setSelectedPicture}
+      selectedItem={setSelectedDish}
     />
   );
 };
 
-export default PicturesDashboard;
+export default DishsDashboard;
